requirements.txt
# GUI Framework
customtkinter==5.2.1
pillow==10.1.0

# Database
sqlalchemy==2.0.23

# API Requests
requests==2.31.0

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2
pandas==2.1.4

# Date/Time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0
